exception handler - delete

+++ src/main/java/com/rodrigues/course/services/UserService.java
@@ -2,10 +2,14 @@ package com.rodrigues.course.services;
 
 import com.rodrigues.course.entities.User;
 import com.rodrigues.course.repositories.UserRepository;
+import com.rodrigues.course.services.exceptions.DatabaseException;
 import com.rodrigues.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Optional;
 
@@ -30,8 +34,15 @@ public class UserService {
         return repository.save(obj);
     }
 
-    public void delete(Long id) {
-        repository.deleteById(id);
+    public void delete(Long id){
+        try {
+            if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        } catch (ResourceNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
@@ -46,6 +57,4 @@ public class UserService {
         entity.setEmail(obj.getEmail());
         entity.setPhone(obj.getPhone());
     }
-
-
 }
